+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - O2V-168: Fix error handling in VHI vm creation
 - O2V-159: Preserve Windows VM hostname
 - O2V-175: Skip migration on onapp vm marked as VIP and fix error handling when migrating existing vm in VHI
+- O2V-172: Fix error handling and pre-checks for migration with placement
 
 ## v1.0.1 (30/06/2023)